Trip_planner/requirements.txt
crewai
python-dotenv
requests
crewai-tools 